moved buttonSlider to button.ts
