+++ src/ui/interface.js
@@ -375,32 +375,6 @@ export class UI {
 		const slider = document.getElementById('sfx');
 		slider.addEventListener('input', () => (game.soundsys.allEffectsMultiplier = slider.value));
 
-		// Prevents default touch behaviour on slider when first touched (prevents scrolling the screen).
-		slider.addEventListener('touchstart', (e) => e.preventDefault(), { passive: false });
-
-		slider.addEventListener('touchmove', (e) =>{
-			// Get slider relative to the view port.
-			const sliderRect = slider.getBoundingClientRect();
-			// The original touch point Y coordinate relative to the view port.
-			const touchPoint = e.touches[0].clientY;
-			/// The y coord of the touch event relative to the slider.
-			const touchRelToSlider = touchPoint - sliderRect.top;
-			const distFromBottom = sliderRect.height - touchRelToSlider;
-			// Normalize the distance from the bottom of the slider to a value between 0 and 1.
-			const normDist = distFromBottom / sliderRect.height;
-			// Scale normDist to range of the slider.
-			const scaledDist = normDist * (slider.max - slider.min);
-			// New value of the slider.
-			const slidersNewVal = scaledDist + parseFloat(slider.min);
-			// Sets the slider value to the new value between the min/max bounds of the slider.
-			slider.value = Math.min(Math.max(slidersNewVal, slider.min), slider.max);
-			
-			// Manually dispatches the input event to update the sound system with new slider value.
-			slider.dispatchEvent(new Event('input'));
-		});
-
-
-
 		this.hotkeys = new Hotkeys(this);
 		const ingameHotkeys = getHotKeys(this.hotkeys);
 
@@ -977,25 +951,6 @@ export class UI {
 
 			if (activeCreature.player.getNbrOfCreatures() > game.creaLimitNbr) {
 				$j('#materialize_button p').text(game.msg.ui.dash.materializeOverload);
-			} 
-			// Check if the player is viewing the wrong tab
-			else if (
-				activeCreature.player.id !== player &&
-				activeCreature.isDarkPriest() &&
-				activeCreature.abilities[3].testRequirements() &&
-				activeCreature.abilities[3].used === false
-			) {
-				$j('#materialize_button p').text(game.msg.ui.dash.wrongPlayer);
-
-				// Switch to turn player's dark priest
-				this.materializeButton.click = () => {
-					this.showCreature("--", activeCreature.player.id);
-				};
-
-				$j('#card .sideA').on('click', this.materializeButton.click);
-				$j('#card .sideA').removeClass('disabled');
-				this.materializeButton.changeState(ButtonStateEnum.glowing);
-				$j('#materialize_button').show();
 			} else if (
 				!summonedOrDead &&
 				activeCreature.player.id === player &&
